requirements.txt
PyQt5>=5.15
opencv-python>=4.5
face_recognition>=1.3.0
# Face recognition models dependency (install via git)
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models
numpy>=1.19
faiss-cpu>=1.7.0
pillow>=8.0